+++ homework/jeon-suwon/re-api/db/todoitem.js
@@ -1,0 +1,19 @@
+const todoData = [
+  {
+    id: 1,
+    userId: 1,
+    title: "할일1",
+    doneAt: "2021-08-01",
+    createdAt: "2021-08-01",
+    updatedAt: "2021-08-01",
+  },
+  {
+    id: 2,
+    userId: 1,
+    title: "할일2",
+    doneAt: "2021-08-01",
+    createdAt: "2021-08-01",
+    updatedAt: "2021-08-01",
+  },
+];
+export { todoData };
