+++ homework/jeon-suwon/re-api/db/user.js
@@ -1,0 +1,13 @@
+const user = [
+  {
+    userId: 1,
+    email: "aaaa1234@naver.com",
+    password: "1234",
+    name: "전수원",
+    role: "student",
+    createdAt: "2021-08-01",
+    updatedAt: "2021-08-01",
+  },
+];
+
+export { user };
